backend/requirements.txt
# ==================== 核心框架 ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ==================== 数据库 ====================
sqlalchemy==2.0.23

# ==================== LLM SDK ====================
zhipuai

# ==================== 缓存和异步 ====================
tenacity==8.2.3

# ==================== 日志 ====================
structlog==23.2.0
python-json-logger==2.0.7

# ==================== 工具库 ====================
python-dotenv==1.0.0
requests>=2.31.0

# ==================== RSS 工具（保留原有功能）====================
feedparser>=6.0.10
tqdm>=4.66.0

# ==================== 测试 ====================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# ==================== 开发工具 ====================
black==23.12.0
flake8==6.1.0
mypy==1.7.1
